pyproject.toml
[project]
name = "namnesis"
version = "2.0.0"
description = "Namnesis - Sovereign AI Agent Protocol"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  # Core (preserved from Resurrectum)
  "argon2-cffi>=23.1.0",
  "boto3>=1.34.0",
  "click>=8.1.0",
  "cryptography>=42.0.0",
  "jsonschema>=4.21.0",
  "rfc8785>=0.1.4",
  # New: On-chain interaction (lightweight, no web3.py)
  "eth-account>=0.10.0",
  "eth-abi>=5.0.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0.0",
]
compression = [
  "py7zr>=0.20.0",
]
all = [
  "py7zr>=0.20.0",
  "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[project.scripts]
namnesis = "namnesis.cli:main"

[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/namnesis"]
